+++ src/pages/[network]/faucet.tsx
@@ -1,3 +1,0 @@
-import Faucet from "lib/pages/faucet";
-
-export default Faucet;